Top Headers Set; ViewModel and ExtensionMethod Modified
I have for, some reason, thought that I was suppost to send to the view a list of elements, whereas I only needed to return the Count of each. I have changed the extension method and viewmodel in accordance to this criterion.

I have set the top headers. Taking a 15 min break here.